Moved mentor wizard components into their own page
